requirements.txt
certifi>=2021.5.30
chardet>=4.0.0
idna>=2.10
Pillow>=8.2.0
pytube>=10.8.4
requests>=2.25.1
urllib3>=1.26.5
